Minor changes for naming, moving classes, etc
